Change Entity from abstract to normal class and also x,y types from float to int

+++ src/main/java/it/unibo/dragondash/model/api/game/Entity.java
@@ -3,13 +3,13 @@ package it.unibo.dragondash.model.api.game;
 import it.unibo.dragondash.view.api.GameView.Sprite;
 
 public class Entity {
-    protected  float x,y;
+    protected  int x,y;
 
     public Sprite sprite;
     
     protected int width, height;
 
-    public Entity(float x, float y, int width, int height) {
+    public Entity(int x, int y, int width, int height) {
         this.x = x;
         this.y = y;
         this.width = width;
